FC backgrounds
### What's done:
* Moved all custom hooks to CustomHooks.kt
* Added custom hook to set background using css styles
