+++ save-frontend/src/main/kotlin/com/saveourtool/save/frontend/components/views/demo/DiktatDemoView.kt
@@ -7,7 +7,9 @@ package com.saveourtool.save.frontend.components.views.demo
 import com.saveourtool.save.frontend.components.basic.cardComponent
 import com.saveourtool.save.frontend.components.basic.diktatDemoComponent
 import com.saveourtool.save.frontend.externals.reactace.AceThemes
+import com.saveourtool.save.frontend.utils.Style
 import com.saveourtool.save.frontend.utils.selectorBuilder
+import com.saveourtool.save.frontend.utils.useBackground
 import com.saveourtool.save.utils.Languages
 
 import csstype.*
@@ -19,6 +21,7 @@ private val backgroundCard = cardComponent(hasBg = true, isPaddingBottomNull = t
 val diktatDemoView = diktatDemoView()
 
 private fun diktatDemoView(): VFC = VFC {
+    useBackground(Style.WHITE)
     val (selectedTheme, setSelectedTheme) = useState(AceThemes.preferredTheme)
     div {
         className = ClassName("d-flex justify-content-center mb-2")
